Moved all utils to utils.js
